refactored properly
